service/requirements.txt
torchaudio
pydub
fastapi
uvicorn
numpy
librosa
soundfile
transformers
python-multipart
https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip

